requirements.txt
mysql-connector-python
pandas
smolagents
huggingface_hub
python-dotenv